requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
gitpython>=3.1.30
matplotlib>=3.3
numpy>=1.23.5
opencv-python>=4.1.1
pillow>=10.3.0
psutil  # system resources
PyYAML>=5.3.1
requests>=2.32.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
torch==2.3.1  # see https://pytorch.org/get-started/locally (recommended)
torchvision==0.18.1
tqdm>=4.64.0
ultralytics>=8.2.34  # https://ultralytics.com
# protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012


# Plotting --------------------------------------------------------------------
pandas
seaborn


# Rest Framework --------------------------------------------------------------
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT